1) Changes for removal of hardcoded service tax. 2) Pdf bill formatting changes done

+++ fuel/modules/cutting_instruction/views/_blocks/layout.php
@@ -184,15 +184,15 @@ function loadfolderlist(account, accname) {
             thisdata["bundlenumber"] = item.bundlenumber;
 			thisdata["processdate"] = item.processdate;
             thisdata["length(in mm)"] = item.length;
-            thisdata["noofsheets"] = item.noofsheets;
+            thisdata["No of sheets"] = item.noofsheets;
             thisdata["weight(in Kgs)"] = item.weight;
-            //var dl = '<a class="ico_coil_delete" title="Delete" href="'+item.dl+'"><img src="<?php echo img_path('iconset/ico_cancel.png'); ?>" /></a>';
-			var dl = '<a class="ico_coil_delete" title="Delete" href="'+item.dl+'" onClick=deleteItem('+item.bundlenumber+')><img src="<?php echo img_path('iconset/ico_cancel.png'); ?>" /></a>';
-			
-            var edit = '<a class="ico_coil_edit" title="Edit" href="#" onClick=radioload('+item.bundlenumber+','+item.processdate+','+item.length+','+item.noofsheets+')><img src="<?php echo img_path('iconset/ico_edit.png'); ?>" /></a>';
-            thisdata["action"] =  edit+' '+dl;
-			//thisdata["action"] = '';
-            partydata.push(thisdata);
+            
+            if( item.status !== 'Billed' && item.status !== 'Ready To Bill') {
+				var dl = '<a class="ico_coil_delete" title="Delete" href="'+item.dl+'" onClick=deleteItem('+item.bundlenumber+')><img src="<?php echo img_path('iconset/ico_cancel.png'); ?>" /></a>';
+				var edit = '<a class="ico_coil_edit" title="Edit" href="#" onClick=radioload('+item.bundlenumber+','+item.processdate+','+item.length+','+item.noofsheets+')><img src="<?php echo img_path('iconset/ico_edit.png'); ?>" /></a>';
+	            thisdata["action"] =  edit+' '+dl;
+        	}
+			partydata.push(thisdata);
 			}
 			if (partydata.length) {
             // If there are files
@@ -328,7 +328,6 @@ function savechange(id){
 	if(parseInt(bundleweight) > parseInt(wei) ){
 		alert('Sorry the Total weight of bundle is more then weight of coil pleae edit the weight to progress!!');
 	}
-		
 	else{
 	$("#check_bar").html('<span style="font-size:20px; color:red">Please wait.. Saving might take some time..</span>');
 	'<i class="icon-refresh icon-spin"></i>';
